chore: move getInstanceMetadataHost to utils folder
