+++ packages/credential-provider-imds/src/utils/getInstanceMetadataHost.spec.ts
@@ -1,10 +1,10 @@
 import { loadConfig } from "@aws-sdk/node-config-provider";
 import { parseUrl } from "@aws-sdk/url-parser";
 
-import { Endpoint } from "./config/Endpoint";
-import { ENDPOINT_CONFIG_OPTIONS } from "./config/EndpointConfigOptions";
-import { EndpointMode } from "./config/EndpointMode";
-import { ENDPOINT_MODE_CONFIG_OPTIONS } from "./config/EndpointModeConfigOptions";
+import { Endpoint } from "../config/Endpoint";
+import { ENDPOINT_CONFIG_OPTIONS } from "../config/EndpointConfigOptions";
+import { EndpointMode } from "../config/EndpointMode";
+import { ENDPOINT_MODE_CONFIG_OPTIONS } from "../config/EndpointModeConfigOptions";
 import { getInstanceMetadataHost } from "./getInstanceMetadataHost";
 
 jest.mock("@aws-sdk/node-config-provider");
